Add files via upload
Class Management System
Almost similar to the School Management System is the Class Management System (CSM). But unlike the School management system which has various aspects to cover like fee management and staff management and so on, I won't be covering those and specifically focus on class management systems which monitors attendance, student information, gradings and maybe basic stuff like courses and schedules. The project may include minute to major details of the student within the class like student ID, contact details, parents information and so on. The project uses SQLite DB for the database and JAVA programming language as a whole project completion.